Add working folder.

+++ ansible/scripts/install_nginx_streaming.sh
@@ -2,10 +2,15 @@
 
 sudo apt-get install -y build-essential libpcre3 libpcre3-dev libssl-dev unzip ffmpeg
 
-mkdir ~/working && cd ~/working
+mkdir -p ~/working && cd ~/working
 
-wget https://nginx.org/download/nginx-1.11.8.tar.gz
-wget -O nginx-rtmp-module.zip https://github.com/arut/nginx-rtmp-module/archive/master.zip
+if [ ! -f ~/working/nginx-1.11.8.tar.gz ]; then
+    wget https://nginx.org/download/nginx-1.11.8.tar.gz
+fi
+
+if [ ! -f ~/working/nginx-rtmp-module.zip ]; then
+    wget -O nginx-rtmp-module.zip https://github.com/arut/nginx-rtmp-module/archive/master.zip
+fi
 
 tar -zxvf nginx-1.11.8.tar.gz
 unzip nginx-rtmp-module.zip
